+++ udata/commands/tests/test_fixtures.py
@@ -51,6 +51,21 @@ class FixturesTest:
             fixtures_fd.flush()
             assert "Fixtures saved to file " in result.output
 
+            # Delete everything, so we can make sure the objects are imported.
+            models.Organization.drop_collection()
+            models.Dataset.drop_collection()
+            models.Discussion.drop_collection()
+            models.CommunityResource.drop_collection()
+            models.User.drop_collection()
+            models.Dataservice.drop_collection()
+
+            assert models.Organization.objects(slug=org.slug).count() == 0
+            assert models.Dataset.objects.count() == 0
+            assert models.Discussion.objects.count() == 0
+            assert models.CommunityResource.objects.count() == 0
+            assert models.User.objects.count() == 0
+            assert models.Dataservice.objects.count() == 0
+
             # Then load them in the database to make sure they're correct.
             result = cli("import-fixtures", fixtures_fd.name)
         assert models.Organization.objects(slug=org.slug).count() > 0
@@ -59,6 +74,8 @@ class FixturesTest:
         assert models.CommunityResource.objects.count() > 0
         assert models.User.objects.count() > 0
         assert models.Dataservice.objects.count() > 0
+        # Make sure we also import the dataservice organization
+        assert models.Dataservice.objects(organization__exists=True).count() > 0
 
     def test_import_fixtures_from_default_file(self, cli):
         """Test importing fixtures from udata.commands.fixture.DEFAULT_FIXTURE_FILE."""